Abm techno (#5)
* deleted unwanted folders,working on mongoose hooks

* Merge branch 'master' into develop

* finished hooks method

* pulled latest

* making chnag

* worked on dealer and employer

* implmented transactions

* renamed Status to EntityStatys

* made changes

---------

+++ .eslintrc.js
@@ -22,7 +22,7 @@ module.exports = {
     '@typescript-eslint/explicit-module-boundary-types': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
     "prettier/prettier": "off",
-    "@typescript-eslint/quotes": ["error", "single", {
+    "@typescript-eslint/quotes": ["error", "double", {
       "avoidEscape": true, "allowTemplateLiterals": true
     }
     ]
